# v6: refactor async code