The async callback-driven code is hard to reason about, [racy](https://github.com/configcat/go-sdk/issues/14),
and requires dynamic type conversions. Also, the concerns between the refresh policies and
the cache seem like they could be better separated.

This PR pushes most of the logic into the config fetcher, leaving the individual policy implementations
as a thin layer on top of that. Externally visible behaviour should remain the same.

This makes things a bit faster too:

```
benchmark               old ns/op     new ns/op     delta
BenchmarkGetValue-8     825           467           -43.32%

benchmark               old allocs     new allocs     delta
BenchmarkGetValue-8     10             6              -40.00%

benchmark               old bytes     new bytes     delta
BenchmarkGetValue-8     504           304           -39.68%
```